move the events into the Filesystem namespace
