+++ core-bundle/src/Filesystem/Dbafs/RetrieveDbafsMetadataEvent.php
@@ -10,7 +10,9 @@ declare(strict_types=1);
  * @license LGPL-3.0-or-later
  */
 
-namespace Contao\CoreBundle\Event;
+namespace Contao\CoreBundle\Filesystem\Dbafs;
+
+use Contao\CoreBundle\Event\AbstractDbafsMetadataEvent;
 
 class RetrieveDbafsMetadataEvent extends AbstractDbafsMetadataEvent
 {
