+++ core-bundle/src/Filesystem/Dbafs/StoreDbafsMetadataEvent.php
@@ -10,7 +10,9 @@ declare(strict_types=1);
  * @license LGPL-3.0-or-later
  */
 
-namespace Contao\CoreBundle\Event;
+namespace Contao\CoreBundle\Filesystem\Dbafs;
+
+use Contao\CoreBundle\Event\AbstractDbafsMetadataEvent;
 
 class StoreDbafsMetadataEvent extends AbstractDbafsMetadataEvent
 {
